requirements.txt
django==2.1.5
Pillow
django-widget-tweaks
django-bootstrap-datepicker-plus
django-bootstrap4
django-debug-toolbar==1.8
django-crispy-forms==1.8.0
gunicorn==19.4.5
psycopg2==2.7.3.1
whitenoise==3.3.1
